+++ server/db/schema/index.ts
@@ -1,4 +1,1 @@
-// Please add drizzle schema.
-// **.ts
-
-export * from "./";
+export * from "./auth";